fix: 修复借助战卡住的问题，修复一个编译错误

+++ src/MeoAssistant/Config/Miscellaneous/OcrPack.cpp
@@ -70,7 +70,7 @@ bool asst::OcrPack::load(const std::filesystem::path& path)
         m_ocr_option->UsePaddleInferBackend();
         break;
     default:
-        Log.error("Unknown OCR Backend", m_backend);
+        Log.error("Unknown OCR Backend", static_cast<int>(m_backend));
         return false;
     }
